add user prompts to main.rb and replace require
